requirements.txt
django
djangorestframework
django-filter
psycopg2-binary
openai
django-cors-headers
